% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_animate_fct_helpers.R
\name{create_metamers}
\alias{create_metamers}
\title{Create a series of metamer data sets}
\usage{
create_metamers(datasets, perturbation = 0.08, N = 30000, trim = 150, ...)
}
\arguments{
\item{datasets}{vector of data sets}

\item{perturbation}{Numeric with the magnitude of the random perturbations.
Can be of length 1 or \code{length(change)}.}

\item{N}{Number of iterations.}

\item{trim}{Max number of metamers to return.}

\item{...}{additional arguments directly to metamer metamerize function}
}
\value{
a metamer list object (list of data frames)
}
\description{
Create a series of metamer data sets
}
