% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_animate_fct_helpers.R
\name{render_animation_graph}
\alias{render_animation_graph}
\title{Create plotly version of animation}
\usage{
render_animation_graph(metamer_df, metamer_sum, frame = 100)
}
\arguments{
\item{metamer_df}{tidy data frame of metamer transition frames}

\item{metamer_sum}{tidy data frame of metamer summary statistics}

\item{frame}{amount of time between frames (in milliseconds)}
}
\value{
plotly object
}
\description{
Create plotly version of animation
}
