% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_explore_fct_helpers.R
\name{render_data_graph}
\alias{render_data_graph}
\title{Render a single datasauRus graph in plotly}
\usage{
render_data_graph(df)
}
\arguments{
\item{df}{data frame corresponding to a single dataset}

\item{...}{additional params that might be used later}
}
\value{
plotly object
}
\description{
Render a single datasauRus graph in plotly
}
